chapter8/backend/requirements.txt
bcrypt
cloudinary
colorama
cryptography
defusedxml
Deprecated
distlib
dnslib
dnspython
email-validator
fastapi
idna
jwcrypto
motor
multidict
passlib
Pillow
pipenv
pycparser
pydantic
PyJWT
pymongo
PySocks
python-dateutil
python-decouple
python-dotenv
python-jwt
python-multipart
PythonDNS
requests
uvicorn
virtualenv
virtualenv-clone